// workspaces/frontend/src/types/route-types.ts
export enum Pages {
  HOME = 'HOME',
  TRANSACTIONS = 'TRANSACTIONS',
  APPROVE = 'APPROVE',
  INVESTIGATE = 'INVESTIGATE',
  FINANCE_SETTLEMENTS = 'FINANCE_SETTLEMENTS',
  FINCANCE_BALANCE = 'FINCANCE_BALANCE',
  USER_ACCOUNTS = 'USER_ACCOUNTS',

  /* TOOLS */
  TOOLS_KYC_SEARCH = 'TOOLS_KYC_SEARCH',
  TOOLS_KYC_BLOCK = 'TOOLS_KYC_BLOCK',
  TOOLS_KYC_ROUTING = 'TOOLS_KYC_ROUTING',
  TOOLS_KYC_FALLBACK = 'TOOLS_KYC_FALLBACK',
  TOOLS_STORE_PRODUCT = 'TOOLS_STORE_PRODUCT',
  TOOLS_STORE_VOUCHER = 'TOOLS_STORE_VOUCHER',

  /* ANALYTICS */
  ANALYTICS_ACCEPTANCE_RATE_OVERVIEW = 'ANALYTICS_ACCEPTANCE_RATE_OVERVIEW',
  ANALYTICS_ACCEPTANCE_RATE_OVERVIEW_MJ = 'ANALYTICS_ACCEPTANCE_RATE_OVERVIEW_MJ',
  ANALYTICS_DRILLDOWN_USERS = 'ANALYTICS_DRILLDOWN_USERS',
  ANALYTICS_GEOGRAPHICAL_OVERVIEW = 'ANALYTICS_GEOGRAPHICAL_OVERVIEW',
  ANALYTICS_USER_OVERVIEW = 'ANALYTICS_USER_OVERVIEW',

  /* RULES */
  RULES_ROUTING_ROUTING = 'RULES_ROUTING_ROUTING',
  RULES_ROUTING_FALLBACK = 'RULES_ROUTING_FALLBACK',
  RULES_ROUTING_FRAUD_PROVIDER = 'RULES_ROUTING_FRAUD_PROVIDER',
  RULES_ROUTING_3DS2_ROUTING = 'RULES_ROUTING_3DS2_ROUTING',
  RULES_ROUTING_3DS2_FALLBACK = 'RULES_ROUTING_3DS2_FALLBACK',
  RULES_FEES_FEE = 'RULES_FEES_FEE',
  RULES_FEES_PROVIDER_FEE = 'RULES_FEES_PROVIDER_FEE',
  RULES_LIMIT = 'RULES_LIMIT',
  RULES_FRAUD_BLOCKS_GENERAL_BLOCK = 'RULES_FRAUD_BLOCKS_GENERAL_BLOCK',
  RULES_FRAUD_BLOCKS_IP_BLOCK = 'RULES_FRAUD_BLOCKS_IP_BLOCK',
  RULES_FRAUD_BLOCKS_BIN_BLOCK = 'RULES_FRAUD_BLOCKS_BIN_BLOCK',
  RULES_FRAUD_BLOCKS_VELOCITY = 'RULES_FRAUD_BLOCKS_VELOCITY',
  RULES_FRAUD_BLOCKS_USER_PSP_ACCOUNT = 'RULES_FRAUD_BLOCKS_USER_PSP_ACCOUNT',
  RULES_FRAUD_BLOCKS_FRAUD_PROVIDER_BLOCK = 'RULES_FRAUD_BLOCKS_FRAUD_PROVIDER_BLOCK',
  RULES_FOREX = 'RULES_FOREX',
  RULES_STATUS_CODE = 'RULES_STATUS_CODE',
  RULES_PAYMENT_METHODS = 'RULES_PAYMENT_METHODS',
  RULES_ALERTS = 'RULES_ALERTS',
  RULES_NOTIFICATION = 'RULES_NOTIFICATION',
  RULES_APPROVAL = 'RULES_APPROVAL',
  RULES_VAT = 'RULES_VAT',
  RULES_VOID_CAPTURE = 'RULES_VOID_CAPTURE',

  /* ADMIN */
  ADMIN_BO_USERS = 'ADMIN_BO_USERS',
  ADMIN_AUDIT_LOG = 'ADMIN_AUDIT_LOG',
  ADMIN_CONFIGURATION = 'ADMIN_CONFIGURATION',
  ADMIN_MERCHANT = 'ADMIN_MERCHANT',
  ADMIN_MAINTENANCE = 'ADMIN_MAINTENANCE',
  ADMIN_API_CLIENTS = 'ADMIN_API_CLIENTS',
  ADMIN_INVOICE = 'ADMIN_INVOICE',
  ADMIN_CUSTOMER = 'ADMIN_CUSTOMER',
  ADMIN_TEMPLATES_EMAIL = 'ADMIN_TEMPLATES_EMAIL',
  ADMIN_TEMPLATES_SCRIPT = 'ADMIN_TEMPLATES_SCRIPT',
  ADMIN_TEMPLATES_HTML = 'ADMIN_TEMPLATES_HTML',
  ADMIN_TEMPLATES_SQL = 'ADMIN_TEMPLATES_SQL',
  ADMIN_TEMPLATES_MQ = 'ADMIN_TEMPLATES_MQ',
  ADMIN_TEMPLATES_CSS = 'ADMIN_TEMPLATES_CSS',
  ADMIN_MESSAGES = 'ADMIN_MESSAGES',
  ADMIN_FOREX = 'ADMIN_FOREX',
  ADMIN_PSP_ICONS = 'ADMIN_PSP_ICONS',
  ADMIN_IIN = 'ADMIN_IIN',
  ADMIN_BANK_MAPPING = 'ADMIN_BANK_MAPPING',
  ADMIN_RESOURCE_CENTER = 'ADMIN_RESOURCE_CENTER',

  // Headers in the menu. We still want to name these in a types way -
  // but they don't get displayed as their own pages
  TOOLS = 'TOOLS',
  KYC = 'KYC',
  STORE = 'STORE',
  FINANCE = 'FINANCE',
  ANALYTICS = 'ANALYTICS',
  RULES = 'RULES',
  ROUTING = 'ROUTING',
  FEES = 'FEES',
  LIMIT = 'LIMIT',
  FRAUD_BLOCKS = 'FRAUD_BLOCKS',
  ADMIN = 'ADMIN',
  TEMPLATES = 'TEMPLATES',
}

export enum RoutePaths {
  HOME = '/',
  TRANSACTIONS = '/transactions',
  APPROVE = '/approve',
  INVESTIGATE = '/investigate',
  FINANCE_SETTLEMENTS = '/finance/settlements',
  FINCANCE_BALANCE = '/finance/balance',
  USER_ACCOUNTS = '/user-accounts',

  /* ANALYTICS */
  ANALYTICS_ACCEPTANCE_RATE_OVERVIEW = '/analytics',
  ANALYTICS_ACCEPTANCE_RATE_OVERVIEW_MJ = '/analytics',
  ANALYTICS_DRILLDOWN_USERS = '/analytics',
  ANALYTICS_GEOGRAPHICAL_OVERVIEW = '/analytics',
  ANALYTICS_USER_OVERVIEW = '/analytics',

  /* TOOLS */
  TOOLS_KYC_SEARCH = '/kyc/search',
  TOOLS_KYC_BLOCK = '/kyc/block',
  TOOLS_KYC_ROUTING = '/kyc/routing',
  TOOLS_KYC_FALLBACK = '/kyc/fallback',
  TOOLS_STORE_PRODUCT = '/product/read',
  TOOLS_STORE_VOUCHER = '/voucher/read',

  /* RULES */
  RULES_ROUTING_ROUTING = '/routing/routing',
  RULES_ROUTING_FALLBACK = '/routing/fallback',
  RULES_ROUTING_FRAUD_PROVIDER = '/routing/fraud-provider',
  RULES_ROUTING_3DS2_ROUTING = '/routing/3ds2-routing',
  RULES_ROUTING_3DS2_FALLBACK = '/routing/3ds2-fallback',
  RULES_FEES_FEE = '/fees/fee',
  RULES_FEES_PROVIDER_FEE = '/fees/provider-fee',
  RULES_LIMIT = '/rules/limit',
  RULES_FRAUD_BLOCKS_GENERAL_BLOCK = '/fraud-block/general-block',
  RULES_FRAUD_BLOCKS_IP_BLOCK = '/fraud-block/ip-block',
  RULES_FRAUD_BLOCKS_BIN_BLOCK = '/fraud-block/bin-block',
  RULES_FRAUD_BLOCKS_VELOCITY = '/fraud-block/velocity',
  RULES_FRAUD_BLOCKS_USER_PSP_ACCOUNT = '/fraud-block/user-psp-account',
  RULES_FRAUD_BLOCKS_FRAUD_PROVIDER_BLOCK = '/fraud-block/fraud-provider-block',
  RULES_FOREX = '/rules/forex',
  RULES_STATUS_CODE = '/rules/status-code',
  RULES_PAYMENT_METHODS = '/rules/payment-methods',
  RULES_ALERTS = '/rules/alerts',
  RULES_NOTIFICATION = '/rules/notification',
  RULES_APPROVAL = '/rules/approval',
  RULES_VAT = '/rules/vat',
  RULES_VOID_CAPTURE = '/rules/void-capture',

  /* ADMIN */
  ADMIN_BO_USERS = '/admin/bo-users',
  ADMIN_AUDIT_LOG = '/admin/audit-log',
  ADMIN_CONFIGURATION = '/admin/configuration',
  ADMIN_MERCHANT = '/admin/merchant',
  ADMIN_MAINTENANCE = '/admin/maintenance',
  ADMIN_API_CLIENTS = '/admin/api-clients',
  ADMIN_INVOICE = '/admin/invoice',
  ADMIN_CUSTOMER = '/admin/customer',
  ADMIN_TEMPLATES_EMAIL = '/templates/email',
  ADMIN_TEMPLATES_SCRIPT = '/templates/script',
  ADMIN_TEMPLATES_HTML = '/templates/html',
  ADMIN_TEMPLATES_SQL = '/templates/sql',
  ADMIN_TEMPLATES_MQ = '/templates/mq',
  ADMIN_TEMPLATES_CSS = '/templates/css',
  ADMIN_MESSAGES = '/admin/messages',
  ADMIN_FOREX = '/admin/forex',
  ADMIN_PSP_ICONS = '/admin/psp-icons',
  ADMIN_IIN = '/admin/iin',
  ADMIN_BANK_MAPPING = '/admin/bank-mapping',
  ADMIN_RESOURCE_CENTER = '/admin/resource-center',
}

export enum LegacyRoutePaths {
  HOME = '/',
  TRANSACTIONS = '/#transaction/search/',
  APPROVE = '/#approve/search/',
  INVESTIGATE = '/#investigate/search/',
  FINANCE_SETTLEMENTS = '/#finance/settlements',
  FINCANCE_BALANCE = '/#finance/balance',
  USER_ACCOUNTS = '/#user-psp-accounts/search/',

  /* ANALYTICS */
  ANALYTICS_ACCEPTANCE_RATE_OVERVIEW = '/#analytics',
  ANALYTICS_ACCEPTANCE_RATE_OVERVIEW_MJ = '/#analytics',
  ANALYTICS_DRILLDOWN_USERS = '/#analytics',
  ANALYTICS_GEOGRAPHICAL_OVERVIEW = '/#analytics',
  ANALYTICS_USER_OVERVIEW = '/#analytics',

  /* TOOLS */
  TOOLS_KYC_SEARCH = '/#kyc/search',
  TOOLS_KYC_BLOCK = '/#T16-decisiontable/read',
  TOOLS_KYC_ROUTING = '/#T15-decisiontable/read',
  TOOLS_KYC_FALLBACK = '/#T22-decisiontable/read',
  TOOLS_STORE_PRODUCT = '/#product/read',
  TOOLS_STORE_VOUCHER = '/#voucher/read/',

  /* RULES */
  RULES_ROUTING_ROUTING = '/#T3-decisiontable/read',
  RULES_ROUTING_FALLBACK = '/#T6-decisiontable/read',
  RULES_ROUTING_FRAUD_PROVIDER = '/#T21-decisiontable/read',
  RULES_ROUTING_3DS2_ROUTING = '/#T28-decisiontable/read',
  RULES_ROUTING_3DS2_FALLBACK = '/#T29-decisiontable/read',
  RULES_FEES_FEE = '/#T4-decisiontable/read',
  RULES_FEES_PROVIDER_FEE = '/#T20-decisiontable/read',
  RULES_LIMIT = '/#T5-decisiontable/read',
  RULES_FRAUD_BLOCKS_GENERAL_BLOCK = '/#T27-decisiontable/read',
  RULES_FRAUD_BLOCKS_IP_BLOCK = '/#T1-decisiontable/read/',
  RULES_FRAUD_BLOCKS_BIN_BLOCK = '/#T2-decisiontable/read',
  RULES_FRAUD_BLOCKS_VELOCITY = '/#T7-decisiontable/read',
  RULES_FRAUD_BLOCKS_USER_PSP_ACCOUNT = '/#T19-decisiontable/read',
  RULES_FRAUD_BLOCKS_FRAUD_PROVIDER_BLOCK = '/#T24-decisiontable/read',
  RULES_FOREX = '/#T8-decisiontable/read',
  RULES_STATUS_CODE = '/#T11-decisiontable/read/',
  RULES_PAYMENT_METHODS = '/#T12-decisiontable/read/',
  RULES_ALERTS = '/#T13-decisiontable/read/',
  RULES_NOTIFICATION = '/#T14-decisiontable/read/',
  RULES_APPROVAL = '/#T18-decisiontable/read/',
  RULES_VAT = '/#T10-decisiontable/read/',
  RULES_VOID_CAPTURE = '/#T26-decisiontable/read/',

  /* ADMIN */
  ADMIN_BO_USERS = '/#user',
  ADMIN_AUDIT_LOG = '/admin/audit-log', // TODO, WHAT IS THE URL?
  ADMIN_CONFIGURATION = '/#config/read/',
  ADMIN_MERCHANT = '/#merchant/search',
  ADMIN_MAINTENANCE = '/#maintenance/search',
  ADMIN_API_CLIENTS = '/#adminapiclient/search',
  ADMIN_INVOICE = '/admin/invoice', // TODO, WHAT IS THE URL?
  ADMIN_CUSTOMER = '/#customer/search',
  ADMIN_TEMPLATES_EMAIL = '/#emailtemplate/read/',
  ADMIN_TEMPLATES_SCRIPT = '/#scripttemplate/read/',
  ADMIN_TEMPLATES_HTML = '/#htmltemplate/read/',
  ADMIN_TEMPLATES_SQL = '/#sqltemplate/read/',
  ADMIN_TEMPLATES_MQ = '/#mqtemplate/read/',
  ADMIN_TEMPLATES_CSS = '/#csstemplate/read/',
  ADMIN_MESSAGES = '/#i18nMessageGrid',
  ADMIN_FOREX = '/#forex',
  ADMIN_IIN = '/#issuerinfo/search',
  ADMIN_BANK_MAPPING = '/#bankmapping/search',
  CREDENTIALS_TEMPLATE = '/#credentialstemplate/read',
  ADMIN_RESOURCE_CENTER = '/#resourceCenter/read',

  // Headers in the menu. We still want to name these in a types way -
  // but they don't get displayed as their own pages.
  // We need them here when we use the keys from one enum in another - Pages and LegacyRoutePaths need to match
  TOOLS = 'TOOLS',
  KYC = 'KYC',
  STORE = 'STORE',
  FINANCE = 'FINANCE',
  ANALYTICS = 'ANALYTICS',
  RULES = 'RULES',
  ROUTING = 'ROUTING',
  FEES = 'FEES',
  LIMIT = 'LIMIT',
  FRAUD_BLOCKS = 'FRAUD_BLOCKS',
  ADMIN = 'ADMIN',
  TEMPLATES = 'TEMPLATES',
}